% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_package.R
\name{cite_package}
\alias{cite_package}
\title{Cite a package in rmarkdown}
\usage{
cite_package(
  package,
  brack = TRUE,
  startText = "",
  endText = "",
  pac_df = "refs",
  pac_cite_file = "package_citations.bib"
)
}
\arguments{
\item{package}{Character. Name of package.}

\item{brack}{Logical. If true, include brackets - '('ref')' - in output.}

\item{startText}{Character. Text to include before reference.}

\item{endText}{Character. Text to include after reference.}

\item{pac_df}{Dataframe of package citations, usually read in via
\code{bib2df::bib2df()}}

\item{pac_cite_file}{Name of file in the project root directory containing
package citations. Usually called \code{package_citations.bib} and created with
\code{knitr::write_bib()}.}
}
\value{
Character string with appropriate markdown formatting to insert the
appropriate reference at that point in the text. If \code{refs} did not exist when
the function is called, \code{refs} is created in the global environment and
if \code{package_citations.bib} did not exist it is created in the project root
directory.
}
\description{
Requires that \code{package_citations.bib} exists in the project root directory
and that the data frame \code{refs} exists
(where \code{refs <- bib2df::bib2df(package_citations.bib)}). These will be
created if \code{refs} does not exist when the function is called.
}
\examples{
library("dplyr")
cite_package("dplyr")
}
