% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_name_map.R
\name{data_name_map}
\alias{data_name_map}
\title{Map points from a data frame with coordinates in columns \code{x} and \code{y}}
\usage{
data_name_map(
  df,
  x = "long",
  y = "lat",
  map_class = NULL,
  polys = NULL,
  polys_col = NULL,
  polys_name = NULL,
  polys_palette = NULL,
  aoi = NULL,
  in_crs = 4283,
  out_crs = 4326
)
}
\arguments{
\item{df}{Dataframe with coordinate columns.}

\item{x}{Character name of column with X coordinates (e.g. "long").}

\item{y}{Character name of column with Y coordinates (e.g. "lat").}

\item{map_class}{Optional character name of column to include in map legend.}

\item{polys}{Simple feature with polygons to map points on.}

\item{polys_col}{Column in \code{polys} to include in map legend.}

\item{polys_name}{Character name to give to \code{polys} legend title.}

\item{polys_palette}{Vector of column colours to use for \code{polys}.}

\item{aoi}{Object on which \code{sf::st_bbox} will return a bounding box for the
map output.}

\item{in_crs}{Coordinate reference system of \code{x} and \code{y} in \code{df}. Can be
given as any coordinate reference system recognised by \code{sf::st_crs} (e.g.
EPSG code).}

\item{out_crs}{Coordinate reference system of resulting map.}
}
\value{
tmap object
}
\description{
Map points from a data frame with coordinates in columns \code{x} and \code{y}
}
