% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary_text.R
\name{data_summary_text}
\alias{data_summary_text}
\title{Produce a histogram of points over time from a data frame of occurrences}
\usage{
data_summary_text(
  df,
  site_cols = c("lat", "long"),
  visit_cols = c("lat", "long", "year"),
  data_name
)
}
\arguments{
\item{df}{Data frame of occurrence records, e.g. as made by envImport. Needs aligned attribute names.}

\item{site_cols}{Character vector. Name of columns defining a site. E.g. lat/long, other coordinates, or raster cell ID.}

\item{visit_cols}{Character vector. Name of columns defining a visit i.e. site plus timestamp. E.g. coordinates/cell plus date/year}

\item{data_name}{How to refer to the dataset, usually a dataset name (e.g "BDBSA", "TERN", "Other"). Only used to paste as text}
}
\value{
Text paragraph/s including formatting for markdown
}
\description{
Produce a histogram of points over time from a data frame of occurrences
}
