% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_kable_extra.R
\name{env_kable_extra}
\alias{env_kable_extra}
\title{Standardise and format report tables}
\usage{
env_kable_extra(data, font_size = 12, width = "100\%", height = NULL, ...)
}
\arguments{
\item{data}{dataframe}

\item{font_size}{passed to the \code{font_size} argument of
\code{kableExtra::kable_sytling()}}

\item{width, height}{passed to the matching argument of
\code{kableExtra::scroll_box()}}

\item{...}{passed to \code{knitr::kable()}}
}
\value{
data formatted appropriately for rmarkdown
}
\description{
Includes \code{knitr::kable()}; \code{kableExtra::kable_sytling()};
and \code{kableExtra::scroll_box()}
}
