% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_bib.R
\name{fix_bib}
\alias{fix_bib}
\title{Make package bibliography, including tweaks for known package issues.}
\usage{
fix_bib(bib_file = "packages.bib", make_key = FALSE, is_package_bib = TRUE)
}
\arguments{
\item{bib_file}{Path to bibliography. If not existing, it will be made.}

\item{make_key}{Logical. Generate a new key via author and year? Defaults to
\code{FALSE}.}

\item{is_package_bib}{Logical. Is this a bibliography of package citations?
Defaults to \code{TRUE.}}
}
\value{
Tweaked \code{bib_file} and data frame of references (from
\code{bib2df::bib2df}).
}
\description{
Fixes sentence case applied to package names via \code{knitr::write_bib} process.
Makes heavy use of \code{bib2df} package.
}
