% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_clean.R
\name{knit_clean}
\alias{knit_clean}
\title{Insert the method descriptions common to most projects}
\usage{
knit_clean(include_specific = TRUE)
}
\arguments{
\item{include_specific}{Logical. Include project-specific cleaning methods?}
}
\value{
Markdown child document.
}
\description{
Describes \code{extent} and \code{grain}; \code{envCleaned} cleaning steps; optionally
project-specific cleaning steps; and, optionally, environmental variables.
}
