% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_generic.R
\name{knit_generic}
\alias{knit_generic}
\title{Insert generic text}
\usage{
knit_generic(section = "scale")
}
\arguments{
\item{section}{Character. Name of generic file to knit, excluding
\code{generic_}. e.g. the default "scale" will knit \code{generic_scale.Rmd}}
}
\value{
Markdown child document
}
\description{
e.g. from generic_env_methods.Rmd; generic_occurrence.Rmd
}
