% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_scripting.R
\name{knit_scripting}
\alias{knit_scripting}
\title{Insert the scripted workflow description common to most projects}
\usage{
knit_scripting(pkg_bib_file)
}
\arguments{
\item{pkg_bib_file}{\emph{Absolute} file path to packages.bib (e.g. in report_prep store). Usually needs additional dots .. when run inside a report.}
}
\value{
Markdown child document.
}
\description{
Describes R/RStudio, git, packages, targets, and reporting. Also introduces \code{envBlah} packages, pulling from packages_bib_file
}
