% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bib.R
\name{make_bib}
\alias{make_bib}
\title{Make package bibliography, including tweaks for known package issues.}
\usage{
make_bib(
  bib_file = "packages.bib",
  make_key = FALSE,
  is_package_bib = TRUE,
  pacs = NULL
)
}
\arguments{
\item{bib_file}{Path to bibliography. Will be created if it doesn't exist.}

\item{make_key}{Logical. Generate a new key using author and year.}

\item{is_package_bib}{Logical. Is this a bibliography of package citations?
Defaults to \code{TRUE.}}

\item{pacs}{Character. Name of any R packages to add to \code{bib_file}.}
}
\value{
Tweaked \code{bib_file} and data frame of references (from
\code{bib2df::bib2df}).
}
\description{
Makes use of both \code{\link[knitr:write_bib]{knitr::write_bib()}} \insertCite{R-knitr,knitr2015,knitr2014}{envReport}
and \code{\link[bib2df:bib2df]{bib2df::bib2df()}} \insertCite{R-bib2df}{envReport}.
}
\examples{
\dontrun{
library(envReport)
bib_file <- tempfile()
make_bib(pacs = .packages())
refs
rm(refs)
unlink(bib_file)
}
}
\references{
\insertAllCited{}
}
