% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cluster_pred_map.R
\name{make_cluster_pred_map}
\alias{make_cluster_pred_map}
\title{Create a map of correctly and incorrectly predicted sites}
\usage{
make_cluster_pred_map(
  this_clust,
  full_pred_df,
  cluster_colours,
  context,
  aoi,
  clust_col = "cluster",
  pred_clust_col = "pred_cluster",
  label_for_correct = "correct",
  x = "long",
  y = "lat",
  xy_crs = 4326
)
}
\arguments{
\item{this_clust}{Name of class.}

\item{full_pred_df}{All sites predicted to class.}

\item{cluster_colours}{Dataframe of classes mapped to colours.}

\item{context}{Character names of columns in full_pred_df defining the context.}

\item{aoi}{sf. Area of interest simple feature.}

\item{clust_col}{Character name of column with the class membership.}

\item{pred_clust_col}{Character name of the column with the predicted class membership.}

\item{label_for_correct}{Label to use in maps for correctly predicted sites.}

\item{x}{Column in \code{full_pred_df} with the x coordinates}

\item{y}{Column in \code{full_pred_df} with the y coordinates}

\item{xy_crs}{Anything interpreted by \code{crs} argument of \code{sf::st_as_sf()} as
valid coordinate system for \code{x} and \code{y}.}
}
\description{
Create a map of correctly and incorrectly predicted sites
}
