% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cluster_pred_plot.R
\name{make_cluster_pred_plots}
\alias{make_cluster_pred_plots}
\title{Plot distribution of probabilities for predicting a class}
\usage{
make_cluster_pred_plots(this_clust, pred_df, cluster_colours)
}
\arguments{
\item{this_clust}{Name of class}

\item{pred_df}{Dataframe of predicted probabilities for each class}

\item{cluster_colours}{Dataframe mapping classes to colours. This is also used
to determine how many other classes were in the clustering.}
}
\value{

}
\description{
Plot distribution of probabilities for predicting a class
}
