% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_data_time_plot.R
\name{occ_data_time_plot}
\alias{occ_data_time_plot}
\title{Produce a histogram of points over time from a data frame of occurrences}
\usage{
occ_data_time_plot(df, year_col = "year", date_format, minyear = NULL)
}
\arguments{
\item{df}{Dataframe with year or date column.}

\item{year_col}{Character. Name of column with date stamp. Can be numeric year, a date class eg posix, or character}

\item{date_format}{Character. Format of year_col, if a character, to be parsed by as.Date. E.g. \%d-\%m-\%Y}

\item{minyear}{Integer. Records before this year will be lumped into one bar labelled 'pre-minyear'. Useful for datasets with very long temporal tails eg gbif}
}
\value{
ggplot2 object
}
\description{
Produce a histogram of points over time from a data frame of occurrences
}
