% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_vis_cell_taxa.R
\name{rec_vis_sit_tax}
\alias{rec_vis_sit_tax}
\title{How many records, visits, sites and taxa in a dataframe}
\usage{
rec_vis_sit_tax(df, site_cols, visit_cols, taxa_col)
}
\arguments{
\item{df}{Dataframe with taxa records.}

\item{site_cols}{Name of columns defining sites.}

\item{visit_cols}{Name of columns defining visits.}

\item{taxa_col}{Name of column containing the taxa.}
}
\value{
One row dataframe
}
\description{
How many records, visits, sites and taxa in a dataframe
}
