% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_of_field.R
\name{source_of_field}
\alias{source_of_field}
\title{Identify original source (field, data source) of a column in a data map}
\usage{
source_of_field(data_map, col)
}
\arguments{
\item{data_map}{Dataframe mapping original field to unified fields}

\item{col}{Which field (column) should be summarised here.}
}
\value{
text string
}
\description{
Create a phrase, for use as text, identifying the original field and data
source for any column in a data map.
}
