% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary_text.R
\name{data_summary_text}
\alias{data_summary_text}
\title{Describe occurrence data}
\usage{
data_summary_text(
  df,
  site_cols = c("lat", "long"),
  visit_cols = c("lat", "long", "year"),
  data_name
)
}
\arguments{
\item{df}{Data frame of occurrence records, e.g. as made by envImport. Needs aligned attribute names.}

\item{site_cols}{Character vector. Name of columns defining a site. E.g. lat/long, other coordinates, or raster cell ID.}

\item{visit_cols}{Character vector. Name of columns defining a visit, usually
site (e.g. coordinates/cell) plus timestamp (e.g. date/year), but could
also include others such as a site name or spatial reliability}

\item{data_name}{How to refer to the dataset, usually a dataset name (e.g "BDBSA", "TERN", "Other"). Only used to paste as text}
}
\value{
Text paragraph/s including formatting for markdown
}
\description{
Build a description of occurrence records including number of records/sites/taxa, earliest and latest records/taxa, and summaries of data fields.
}
