% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_changes.R
\name{knit_changes}
\alias{knit_changes}
\title{Create markdown text regarding changes to a workflow}
\usage{
knit_changes(
  project = basename(here::here()),
  include_packages = NULL,
  add_dots = ".",
  use_units = "years",
  suffix_text =
    "The workflow and associated package(s) are undergoing constant improvement."
)
}
\arguments{
\item{project}{Character. Rstudio 'project'}

\item{include_packages}{Character. Any packages to include in the summary
text}

\item{add_dots}{Character. Usually \code{"."} for interactive use or \code{".."} inside
a knit.}

\item{use_units}{Character. Passed to the \code{unit} argument of
\code{lubridate::time_length()}}

\item{suffix_text}{Character. Text to add on to the end of the workflow
summary information}
}
\value{
Text formatted for inclusion in rmarkdown.
}
\description{
Useful in the context of the workflow undergoing constant improvement.
}
