% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_scripting.R
\name{knit_scripting}
\alias{knit_scripting}
\title{Insert the scripted workflow description common to most projects}
\usage{
knit_scripting(
  packages = sort(unique(unname(unlist(yaml::read_yaml(here::here("settings",
    "packages.yaml"))))))
)
}
\arguments{
\item{packages}{Character. Vector of any loaded packages (doesn't have to
be limited to just \code{envBlah} packages).}
}
\value{
Markdown child document.
}
\description{
Describes R/RStudio, git, packages, targets, and reporting. Also introduces
\code{envBlah} packages, pulling from packages_bib_file
}
