% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_of_field.R
\name{source_of_field}
\alias{source_of_field}
\title{Identify original source (field, data source) of a column in a data map}
\usage{
source_of_field(
  data_map,
  this_col,
  first_data_col = which(names(data_map) == "bdbsa")
)
}
\arguments{
\item{data_map}{Dataframe mapping original field to unified fields}

\item{this_col}{Which field (column) should be summarised here.}

\item{first_data_col}{First column in data_map containing data_names.}
}
\value{
text string
}
\description{
Create a phrase, for use as text, identifying the original field and data
source for any column in a data map.
}
